+++ backend/user/apps.py
@@ -4,7 +4,3 @@ from django.apps import AppConfig
 
 class UserConfig(AppConfig):
     name = "user"
-
-    def ready(self) -> None:
-        # Application
-        import user.signals  # noqa: F401